Allow multiple DPUs
Experimenting with asynchronous operation of independent DPU sets.
DPUs are meant to be allocated in groups, but I only one want at a
time. Each DPU operates completely independently, so there is no
good reason to allocate them in groups. I want each DPU to process
its own file, which means there is no reason why the first DPU
should delay its processing while I'm preparing the input buffer
for the second DPU, and so on. Also, if one DPU completes quickly
(i.e. has a short input file) it should be able to receive a new
file ASAP, and not have to wait for batching or ordering of DPUs.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>