Uncomment WRAM variables
I had a problem copying variables from the host to WRAM when the DPUs
were accessed inside a nested loop - DPU_FOREACH() inside DPU_RANK_FOREACH().

Vincent upgraded the SDK to 2020.3.0-unstable and it fixed the problem
with the WRAM copies. However, I am still seeing issues with MRAM copies.
The transfer fails with: (mram is busy). I don't know why the mram is
busy, but I didn't tell it to do anything!

Signed-off-by: Joel Nider <joel@ece.ubc.ca>