Fix input algorithm from stdin
When reading a list of filenames from stdin, we read a buffer one chunk at a time
and parse it, looking for file names. After there is no more data left to read
from stdin, the buffer still has some residual data that needs to be parsed.

This fixes the problem of the 7 missing files!

Signed-off-by: Joel Nider <joel@ece.ubc.ca>