Halt the CPU and wait for debugger
If a DPU is in a 'fault' state and the program has been compiled in
debug mode (DEBUG_DPU=1) enter an infinite loop to give a chance to
connect with dpu_lldb from a different shell.

The basic steps to debug are:
1) find the pid of the host program (ps -ef | grep host)
2) attach the debugger to that pid (dpu-lldb -p <pid>)
3) attach lldb to the specific dpu (dpu_attach <rank>.<slice>.<dpu>)

Then you can inspect the threads, stack frames, variables, etc.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>