Artificially limit the number of input files
This is great for testing with different size datasets, because we
can use the same input directory, and choose the number of files
that will be searched.

This is not so great for comparing with the host 'grep' which I
don't think has the same functionality (at least I haven't found
out how to do it yet).

Signed-off-by: Joel Nider <joel@ece.ubc.ca>