clear the DPU heap when starting
I was stuck on a problem for about 2 weeks, and just figured it out (with Romaric's help). I am calling seqread_init() (to initialize the sequential reader) on the same DPU many times (like about 50) - because I am using the DPUs like a work pool. It worked great for about 10000 files, then I was seeing a crash. It looked like seqread_init was trying to allocate 65K out of WRAM (which is obviously not possible or valid).

The reason: seqread_init() calls mem_alloc(), which takes memory from the DPU heap. The heap pointer does not get initialized when the DPU starts, so it kept allocating more and more memory until eventually I ran out of heap space. There is no warning and no (easy) way to know this is happening.

The solution: call mem_reset() when the DPU starts - it resets the heap pointer, and everything is ok again

Signed-off-by: Joel Nider <joel@ece.ubc.ca>