Define the input memory in terms of the output memory
The input memory should be as large as possible, excluding the memory
reserved for output. Currently, the output memory is 1 MB, and so
the input memory should be the rest of MRAM (currently 32 MB). If we
change the size of MRAM, the input size should change, not the output
size.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>