One performance counter per DPU
It doesn't work with one counter per tasklet - there is only one
shared register!

The counter is started with tasklet 0, assuming that is the first
tasklet to run. The final value is read once per tasklet, but the
final tasklet to complete is the last one to update the variable,
which means we don't need to wait for all of the tasks to finish.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>